+++ src/Enums/VoiceType.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace AhmadMayahi\Polly\Enums;
+
+enum VoiceType: string
+{
+    case Auto = 'auto';
+
+    case Standard = 'standard';
+
+    case Neural = 'neural';
+}