+++ CHANGELOG.md
@@ -2,11 +2,16 @@
 
 All notable changes will be documented in this file.
 
+## 0.3 - 2021-12-30
+
+- Bug fixes.
+- Add `voiceType` method in voices.
+
 ## 0.2 - 2021-12-30
 
-- Bug fixes
-- Add `describe` method in voices
+- Bug fixes.
+- Add `describe` method in voices.
 
 ## 0.1 - 2021-12-25
 
-- initial release
+- initial release.